Изменил пакет dto на entity и изменил классы Address и People
Подключил lombok и аннотировал классы и поля аннотациями Hibernate
